CAN-16: add check for missing template
- checks if template is missing and creates a default if it is

+++ justfile
@@ -141,7 +141,12 @@ generate-cicd TARGET_DIR FLAG="":
     mkdir -p ./generate/.output
     envsubst < generate/config-template.json > generate/.config.json
     cp ./generate/.openapi-generator-ignore ./generate/.output/.openapi-generator-ignore
-    cp ./generate/templates/description.{{APPLICATION_NAME}}.mustache ./generate/templates/description.mustache
+    # Basic check if file exists
+    if [ -f "./generate/templates/description.{{APPLICATION_NAME}}.mustache" ]; then
+        cp ./generate/templates/description.{{APPLICATION_NAME}}.mustache ./generate/templates/description.mustache
+    else
+        echo "FINBOURNE Technology candela-python-sdk" > ./generate/templates/description.mustache;
+    fi
 
     ./generate/generate.sh ./generate ./generate/.output {{swagger_path}} .config.json
     rm -f generate/.output/.openapi-generator-ignore
